Refactor Tx message types
