+++ packages/scrapper/src/database/RegisterAccount.ts
@@ -1,0 +1,27 @@
+import { AccountNft, RegisterAccountValue, StarnameSchemaModel } from "@starname-explorer/shared";
+
+import { StarnameExtension } from "../starname";
+
+export async function RegisterAccount(
+  client: StarnameExtension,
+  account: RegisterAccountValue,
+): Promise<void> {
+  const accountDetails = await client.starname.queryResolve(`${account.name}*${account.domain}`);
+
+  await saveAccountNft(accountDetails);
+}
+
+export async function saveAccountNft(account: AccountNft): Promise<void> {
+  await StarnameSchemaModel.updateOne(
+    { domain: account.domain, name: account.name },
+    account,
+    {
+      upsert: true,
+    },
+    (error) => {
+      if (error) {
+        console.error(error);
+      }
+    },
+  );
+}
