+++ packages/shared/src/index.ts
@@ -1,2 +1,4 @@
+export * from "./messages";
 export * from "./models/Domain/DomainSchema";
 export * from "./models/Starname/StarnameSchema";
+export * from "./types";
