+++ packages/shared/src/messages/DeleteDomain.ts
@@ -1,7 +1,6 @@
 import { Msg } from "@cosmjs/launchpad";
-import { DomainSchemaModel, StarnameSchemaModel } from "@starname-explorer/shared";
 
-interface DeleteDomainValue {
+export interface DeleteDomainValue {
   readonly domain: string;
   /** Bech32 owner address */
   readonly owner: string;
@@ -16,9 +15,4 @@ export interface MsgDeleteDomain extends Msg {
 
 export function isMsgDeleteDomain(msg: Msg): msg is MsgDeleteDomain {
   return (msg as MsgDeleteDomain).type === "starname/DeleteDomain";
-}
-
-export async function MsgDeleteDomainStore(domain: DeleteDomainValue): Promise<void> {
-  await DomainSchemaModel.deleteOne({ domain: domain.domain });
-  await StarnameSchemaModel.deleteOne({ domain: domain.domain });
 }
