+++ packages/shared/src/messages/index.ts
@@ -1,0 +1,10 @@
+export * from "./DeleteAccount";
+export * from "./DeleteDomain";
+export * from "./RegisterAccount";
+export * from "./RegisterDomain";
+export * from "./RenewAccount";
+export * from "./RenewDomain";
+export * from "./ReplaceAccountResources";
+export * from "./SetAccountMetadata";
+export * from "./TransferAccount";
+export * from "./TransferDomain";
