+++ packages/shared/src/types/DomainNft.ts
@@ -1,0 +1,11 @@
+import { DomainType } from "../models/Domain/DomainSchema";
+
+export interface DomainNft {
+  readonly name: string;
+  /** Bech32 account address */
+  readonly admin: string;
+  /** Bech32 account address */
+  readonly broker: string;
+  readonly type: DomainType;
+  readonly valid_until: number;
+}
